sidebar update
