--- digital-store-project/src/components/Logo/app-sidebar.jsx
@@ -1,11 +1,0 @@
-import { Sidebar, SidebarContent } from "@/components/ui/sidebar"
-
-const AppSidebar = () => {
-  return (
-    <Sidebar>
-      <SidebarContent />
-    </Sidebar>
-  )
-}
-
-export default AppSidebar;
